requirements.txt
Scrapy==2.11.1
playwright==1.43.0
scrapy-playwright==0.0.34
fastapi==0.108.0
pydantic==2.5.3
boto3==1.34.100
botocore==1.34.100
python-dotenv~=1.0.0
itemadapter~=0.9.0